refactor(sdk-frontend): move content-entry resolution into WbSdk.getPage()
Instead of requiring callers to invoke `resolveAutoLoad(page, components)`
separately, `WbSdk.getPage()` now accepts an optional `{ components }`
parameter and resolves content-entry inputs internally. This simplifies
the user-facing API and keeps the resolution step co-located with page
fetching.

- Rename resolveAutoLoad → resolveContentEntries (internal, no export)
- Add `GetPageOptions` type with optional `components` array
- Remove `resolveAutoLoad` and `ManifestCarrier` from public exports
